requirements3.txt
asgiref==3.3.1
awscli==1.11.13
blinker==1.3
boto3==1.17.25
botocore==1.20.25
chardet==2.3.0
cloud-init==0.7.9
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
Django==3.1.7
docutils==0.12
idna==2.0
Jinja2==2.8
jmespath==0.10.0
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
MarkupSafe==0.23
mysqlclient==2.0.3
oauthlib==1.0.3
Pillow==8.1.2
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1
python-dateutil==2.8.1
python-debian==0.1.27
python-systemd==231
pytz==2021.1
PyYAML==3.11
requests==2.9.1
roman==2.0.0
rsa==3.2.3
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.26.3
virtualenv==15.0.1
